GTA.xml: fix "Why no Radio" command
In command "Why no Radio" in GTA.xml the word "where" is misplaced in
the second group -- it is already present in the first group.  Update
the command so that first group is only "asking" words, second group is
only verbs (plus "no"), and third group is only synonyms for "radio".

In first group -- question mark isn't used anywhere else as a <word>,
which is suspicious, so let's just drop it.  In second group -- mark
the very short word "no" with `wholeword="true"` and add more verbs to
cover more possible phrasings, like:

  - where is the music?
  - why doesn't the radio play?

In the response, add the missing bit from VCS's copy.