Generic.xml: add "JDQ schedule" command
People tried to trigger Botimuz using "bot schedule" on the first day of
JDQ the Third.  But the "WHEN JDQ" command is composed in a way to
trigger on "bot jdq" and "schedule jdq", so making it trigger on "bot
schedule" is not feasible.  So just introduce a new command just for
"Botimuz schedule" and "bot schedule".  Maybe it could be expanded in
the future.